pyproject.toml
[tool.poetry]
name = "thebitbybit-recruitment"
version = "0.1.0"
description = "Recruitment task"
authors = ["Filip <mularskif@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
django-environ = "^0.10.0"
psycopg2 = "^2.9.6"
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
drf-yasg = "^1.21.5"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"
faker = "^18.5.1"
pytest-cov = "^4.0.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
